refactor: place createStoryProxy funciton and service in same file altogether
